add: chatroom.service
