+++ backend/src/chat/chat.controller.ts
@@ -2,6 +2,7 @@ import { Query, Controller, Get, ParseIntPipe } from '@nestjs/common';
 import { ChatService } from './chat.service';
 import { BanService } from './ban.service';
 import { MuteService } from './mute.service';
+import { ChatroomService } from './chatroom.service';
 import type { ChatUser, ChatMessage } from './types/chat';
 
 @Controller('chat')
@@ -10,6 +11,7 @@ export class ChatController {
     private readonly chatService: ChatService,
     private readonly banService: BanService,
     private readonly muteService: MuteService,
+    private readonly chatroomService: ChatroomService,
   ) {}
 
   /**
@@ -27,7 +29,7 @@ export class ChatController {
     const adminUsers = await this.chatService.findAdmins(roomId);
     const bannedUsers = await this.banService.findBannedUsers(roomId);
     const mutedUsers = await this.muteService.findMutedUsers(roomId);
-    const chatroomOwner = await this.chatService.findChatroomOwner(roomId);
+    const chatroomOwner = await this.chatroomService.findChatroomOwner(roomId);
 
     const adminIds = adminUsers.map((admin) => admin.userId);
     const bannedIds = bannedUsers.map((user) => user.id);
@@ -44,7 +46,7 @@ export class ChatController {
 
     // すべてを満たさないUser一覧を取得する
     const canSetAdminUsers =
-      await this.chatService.findChatroomMembersAsChatUsers({
+      await this.chatroomService.findChatroomMembersAsChatUsers({
         where: {
           chatroomId: roomId,
           userId: {
@@ -69,7 +71,7 @@ export class ChatController {
   ): Promise<ChatUser[]> {
     const bannedUsers = await this.banService.findBannedUsers(roomId);
     const mutedUsers = await this.muteService.findMutedUsers(roomId);
-    const chatroomOwner = await this.chatService.findChatroomOwner(roomId);
+    const chatroomOwner = await this.chatroomService.findChatroomOwner(roomId);
 
     const bannedIds = bannedUsers.map((user) => user.id);
     const mutedIds = mutedUsers.map((user) => user.id);
@@ -80,7 +82,7 @@ export class ChatController {
 
     // すべてを満たさないUser一覧を取得する
     const canSetOwnerUsers =
-      await this.chatService.findChatroomMembersAsChatUsers({
+      await this.chatroomService.findChatroomMembersAsChatUsers({
         where: {
           chatroomId: roomId,
           userId: {
@@ -111,13 +113,12 @@ export class ChatController {
   async findNotMutedUsers(
     @Query('roomId', ParseIntPipe) roomId: number,
   ): Promise<ChatUser[]> {
-    const chatroomUsers = await this.chatService.findChatroomMembersAsChatUsers(
-      {
+    const chatroomUsers =
+      await this.chatroomService.findChatroomMembersAsChatUsers({
         where: {
           chatroomId: roomId,
         },
-      },
-    );
+      });
     const mutedUsers = await this.muteService.findMutedUsers(roomId);
     if (mutedUsers.length === 0) {
       return chatroomUsers;
@@ -150,13 +151,12 @@ export class ChatController {
   async findNotBannedUsers(
     @Query('roomId', ParseIntPipe) roomId: number,
   ): Promise<ChatUser[]> {
-    const chatroomUsers = await this.chatService.findChatroomMembersAsChatUsers(
-      {
+    const chatroomUsers =
+      await this.chatroomService.findChatroomMembersAsChatUsers({
         where: {
           chatroomId: roomId,
         },
-      },
-    );
+      });
 
     const bannedUsers = await this.banService.findBannedUsers(roomId);
     if (bannedUsers.length === 0) {
