+++ backend/src/chat/chat.gateway.ts
@@ -13,6 +13,7 @@ import type { ChatMessage } from './types/chat';
 import { ChatService } from './chat.service';
 import { BanService } from './ban.service';
 import { MuteService } from './mute.service';
+import { ChatroomService } from './chatroom.service';
 import { CreateAdminDto } from './dto/admin/create-admin.dto';
 import { DeleteAdminDto } from './dto/admin/delete-admin.dto';
 import { IsAdminDto } from './dto/admin/is-admin.dto';
@@ -54,6 +55,7 @@ export class ChatGateway {
     private readonly chatService: ChatService,
     private readonly banService: BanService,
     private readonly muteService: MuteService,
+    private readonly chatroomService: ChatroomService,
   ) {}
 
   @WebSocketServer() server: Server;
@@ -102,7 +104,7 @@ export class ChatGateway {
     this.logger.log(`chat:createAndJoinRoom: ${dto.name}`);
 
     // チャットルームを作成する
-    const createdRoom = await this.chatService.createRoom(dto);
+    const createdRoom = await this.chatroomService.create(dto);
     if (!createdRoom) {
       return { createdRoom: undefined };
     }
@@ -135,7 +137,7 @@ export class ChatGateway {
   ) {
     this.logger.log(`chat:getJoinedRooms: ${dto.userId}`);
     // ユーザーが入室しているチャットルームを取得する
-    const rooms = await this.chatService.findJoinedRooms(dto.userId);
+    const rooms = await this.chatroomService.findJoinedRooms(dto.userId);
 
     const clientChatrooms = rooms.map((room) =>
       this.convertToClientChatroom(room),
@@ -286,7 +288,7 @@ export class ChatGateway {
   ): Promise<{ joinedRoom: ClientChatroom | undefined }> {
     this.logger.log(`chat:joinRoom received -> ${dto.userId}`);
 
-    const joinedRoom = await this.chatService.joinRoom(dto);
+    const joinedRoom = await this.chatroomService.joinRoom(dto);
     if (!joinedRoom) {
       return { joinedRoom: undefined };
     }
@@ -321,10 +323,10 @@ export class ChatGateway {
    * @param client
    * @param DeleteChatroomDto
    */
-  @SubscribeMessage('chat:deleteRoom')
+  @SubscribeMessage('chat:delete')
   async deleteRoom(@MessageBody() dto: DeleteChatroomDto): Promise<boolean> {
-    this.logger.log(`chat:deleteRoom received -> roomId: ${dto.id}`);
-    const deletedRoom = await this.chatService.deleteRoom(dto);
+    this.logger.log(`chat:delete received -> roomId: ${dto.id}`);
+    const deletedRoom = await this.chatroomService.delete(dto);
     if (!deletedRoom) {
       return false;
     }
@@ -332,7 +334,7 @@ export class ChatGateway {
     const socketRoomName = this.generateSocketChatRoomName(deletedRoom.id);
 
     const deletedClientRoom = this.convertToClientChatroom(deletedRoom);
-    this.server.to(socketRoomName).emit('chat:deleteRoom', deletedClientRoom);
+    this.server.to(socketRoomName).emit('chat:delete', deletedClientRoom);
 
     // 入室者をルームから退出させる
     this.server.socketsLeave(socketRoomName);
@@ -352,7 +354,7 @@ export class ChatGateway {
   ): Promise<boolean> {
     this.logger.log(`chat:leaveRoom received userId -> ${dto.userId}`);
 
-    const deletedMember = await this.chatService.removeChatroomMember(dto);
+    const deletedMember = await this.chatroomService.removeChatroomMember(dto);
     if (!deletedMember) {
       return false;
     }
@@ -377,7 +379,7 @@ export class ChatGateway {
     const excludeIdSets = new Set([...bannedIds, ...mutedIds]);
     const excludeIds = [...excludeIdSets];
     const chatroomMembers =
-      await this.chatService.findChatroomMembersAsChatUsers({
+      await this.chatroomService.findChatroomMembersAsChatUsers({
         where: {
           chatroomId: dto.chatroomId,
           userId: {
@@ -394,7 +396,7 @@ export class ChatGateway {
       userId: dto.userId,
     };
 
-    const deletedRoom = await this.deleteRoom(deleteChatroomDto);
+    const deletedRoom = await this.chatroomService.delete(deleteChatroomDto);
     if (!deletedRoom) {
       this.logger.log('chat:leaveRoom failed to delete room');
 
@@ -438,7 +440,7 @@ export class ChatGateway {
     const hiddenChatRooms = [ChatroomType.PRIVATE, ChatroomType.DM];
 
     // public, protectedのチャットルーム一覧を取得する
-    const viewableRooms = await this.chatService.findAll({
+    const viewableRooms = await this.chatroomService.findAll({
       where: {
         NOT: {
           type: { in: hiddenChatRooms },
@@ -447,7 +449,7 @@ export class ChatGateway {
     });
 
     // userが所属しているチャットルームの一覧を取得する
-    const joinedRooms = await this.chatService.findJoinedRooms(dto.userId);
+    const joinedRooms = await this.chatroomService.findJoinedRooms(dto.userId);
 
     const roomsDiff = this.getChatroomDiff(viewableRooms, joinedRooms);
     const viewableAndNotJoinedRooms = viewableRooms.filter((item) => {
@@ -520,7 +522,7 @@ export class ChatGateway {
       `chat:updatePassword received -> roomId: ${dto.chatroomId}`,
     );
 
-    return await this.chatService.updatePassword(dto);
+    return await this.chatroomService.updatePassword(dto);
   }
 
   /**
@@ -599,7 +601,7 @@ export class ChatGateway {
   ): Promise<{ chatroom: ClientChatroom | undefined }> {
     this.logger.log('chat:directMessage received');
 
-    const existingDMRoom = await this.chatService.findExistingDMRoom(
+    const existingDMRoom = await this.chatroomService.findExistingDMRoom(
       dto.senderId,
       dto.recipientId,
     );
@@ -617,9 +619,9 @@ export class ChatGateway {
       type: ChatroomType.DM,
       ownerId: dto.senderId,
     };
-    const createdRoom = await this.chatService.createRoom(createChatroomDto);
+    const createdRoom = await this.chatroomService.create(createChatroomDto);
     if (!createdRoom) {
-      this.logger.log('chat:directMessage failed to createRoom');
+      this.logger.log('chat:directMessage failed to create');
 
       return { chatroom: undefined };
     }
@@ -650,7 +652,7 @@ export class ChatGateway {
       this.logger.log(error);
 
       // どちらかの入室処理に失敗した場合は、チャットルームを削除する
-      await this.chatService.deleteRoom({
+      await this.chatroomService.delete({
         id: createdRoom.id,
         userId: dto.senderId,
       });
@@ -678,7 +680,7 @@ export class ChatGateway {
   ): Promise<boolean> {
     this.logger.log('chat:changeRoomOwner received');
 
-    const changedRoom = await this.chatService.updateRoom({
+    const changedRoom = await this.chatroomService.update({
       data: {
         owner: {
           connect: {
@@ -805,7 +807,7 @@ export class ChatGateway {
     await client.join(userRoomName);
 
     // すでに入室中のチャットルームも通知を受け取れるようにする
-    const rooms = await this.chatService.findJoinedRooms(userId);
+    const rooms = await this.chatroomService.findJoinedRooms(userId);
     rooms.map(async (room) => {
       await client.join(this.generateSocketChatRoomName(room.id));
     });
