+++ backend/src/chat/chat.module.ts
@@ -6,10 +6,18 @@ import { UserModule } from '../user/user.module';
 import { PrismaService } from '../prisma/prisma.service';
 import { BanService } from './ban.service';
 import { MuteService } from './mute.service';
+import { ChatroomService } from './chatroom.service';
 
 @Module({
   imports: [UserModule],
   controllers: [ChatController],
-  providers: [ChatGateway, ChatService, PrismaService, BanService, MuteService],
+  providers: [
+    ChatGateway,
+    ChatService,
+    PrismaService,
+    BanService,
+    MuteService,
+    ChatroomService,
+  ],
 })
 export class ChatModule {}
